Rectangle recovered

+++ src/Circle.java
@@ -1,14 +1,14 @@
 public class Circle extends Figure{
 
-    public Circle(double radius){
-        super(new double[]{radius});
+    public Circle(int radius){
+        super(new int[]{radius});
     }
 
-    public double getPerimeter(){
-        return 2 * 3.14 * sides[0];
+    public int getPerimeter(){
+        return (int) (2 * 3.14 * sides[0]);
     }
 
-    public double getArea(){
-        return 3.14 * sides[0] * sides[0];
+    public int getArea(){
+        return (int) (3.14 * sides[0] * sides[0]);
     }
 }
