+++ src/Triangle.java
@@ -1,11 +1,11 @@
 public class Triangle extends Figure{
 
-   public Triangle(double a, double b, double c) {
-       super(new double[]{a,b,c});
+   public Triangle(int a, int b, int c) {
+       super(new int[]{a,b,c});
    }
 
-   public double getArea(){
-       double p = (sides[0] + sides[1] + sides[2]) / 2;
-       return Math.sqrt(p * (p-sides[0]) * (p-sides[1]) * (p-sides[2]));
+   public int getArea(){
+       int p = (sides[0] + sides[1] + sides[2]) / 2;
+       return (int) Math.sqrt(p * (p-sides[0]) * (p-sides[1]) * (p-sides[2]));
    }
 }
